01-Raw-Image-Processing/requirements.txt
numpy>=1.20.0
scipy>=1.7.0
pillow>=9.0.0
